Add island removal functionality
Added the capability to remove islands when clicking on them. The `removeIsland` method in the World class handles the removal process by both updating the internal islands array and removing the corresponding DOM element. The feature enhances the user interaction by allowing islands to be dynamically removed from the world.